Initial commit, version 0.1
* Multicolor, dimmable with multi pwm and fully homekit integrated without bridge

* Originally, this software is partially based on an example, which is
under MIT license. This License was included, as well as a new MIT
for all parts of the software written by me which I'm uncertain whether
they are MIT under the old license or not. Adding my own MIT is only for
clarification purposes! In no way am I claiming ownership over the code not produced
by me or over code of which I must not take ownership due to licensing of original source.